refactor to new prms_runoff_ag.py
