+++ pywatershed/__init__.py
@@ -38,6 +38,7 @@ from .hydrology.prms_hydraulic_geometry import (
     PRMSHydraulicGeometryWidthOnly,
 )
 from .hydrology.prms_runoff import PRMSRunoff
+from .hydrology.prms_runoff_ag import PRMSRunoffAg
 from .hydrology.prms_runoff_no_dprst import PRMSRunoffNoDprst
 from .hydrology.prms_snow import PRMSSnow
 from .hydrology.prms_soilzone import PRMSSoilzone
