+++ pywatershed/hydrology/__init__.py
@@ -16,6 +16,7 @@ from .prms_hydraulic_geometry import (
     PRMSHydraulicGeometryWidthOnly,
 )
 from .prms_runoff import PRMSRunoff
+from .prms_runoff_ag import PRMSRunoffAg
 from .prms_runoff_no_dprst import PRMSRunoffNoDprst
 from .prms_snow import PRMSSnow
 from .prms_soilzone import PRMSSoilzone
